Template Design Refactoring
1. When using TDF using abstract class, where the child classes have same functionality, to which their common functionality can be pulled up to the abstract parent class, if the method can't be pulled up use the lambada expressions with the functional interface - Supplier<Type>, Consumer<Type>

So, we see that 2 methods initializeParameters() and matches() have been abstract and implemented in their child class, initially we didn't refactor the initialize() into read() method,  but we saw the some statements in initialize() method were identical, so we extracted the statements that were different and pulled up rest of code that was similar to parent class. And same thing we did we the matches() function, we implemented the different part into the child class and pulled the rest into parent abstract class.

And there was problem what if we have parent class that is not abstract or the there are objects created from the parent class, in that case we need to go lambda functions for which we have Supplier<> and Consumer<> interface and many more in java doc website for further reference.