polymorphism Usage - where there is need for If and else type checking
So, we 5 classes, and we see that CallEdge class is using polymorphism, so at run time it decide as per the type of endNode variable which of its implementation be called either from its child classes or its parent abstract class.

See from the callNode.png, earlier it was like that in image, but we reduced it down to just 1 line with the help of polymorphism.